requirements.txt
absl-py==0.9.0
aio-pika==6.4.1
aiormq==3.2.2
astor==0.8.1
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==3.0
cymem==2.0.3
Cython==0.29.14
DAWG-Python==0.7.2
deeppavlov==0.11.0
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
fastapi==0.47.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.30.0
h11==0.9.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.7.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
multidict==4.7.6
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.0
opt-einsum==3.3.0
overrides==2.7.0
pamqp==2.3.0
pandas==0.25.3
plac==1.1.3
preshed==3.0.2
protobuf==3.12.2
pycparser==2.20
pydantic==1.3
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.404381.4453942
pyOpenSSL==19.1.0
pyTelegramBotAPI==3.6.7
python-dateutil==2.8.1
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.15.100
rusenttokenize==0.0.5
sacremoses==0.0.35
scikit-learn==0.21.2
scipy==1.4.1
six==1.15.0
spacy==2.2.3
srsly==1.0.2
starlette==0.12.9
tensorboard==1.15.0
tensorflow-gpu==1.10.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
thinc==7.3.1
tqdm==4.41.1
urllib3==1.25.10
uvicorn==0.11.1
wasabi==0.7.1
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
